o.c.swt/opibuilder.widgets: Changed Linking Containers resizing
When a linking container contains an OPI which is not the same size
as it, there are different ways of handling this. These have been
changed from a pair of boolean values (auto_size, zoom_to_fit) to
a combo property (resize_behaviour).

The neccassary changes to update older OPIs to the new behaviour
have also been made.