o.c.util.clock: Replace ViewPart with E4 Pojo
Example for 'e4view' that allows using E4 Pojo in RCP:
Part is now Pojo, using DI to get parent and preference setting.
Could have used the same ID as the View, but that creates problems when
existing workspace had already used View with that ID.
Command to open the view, and Preference page remain unchanged.