org.csstudio.opibuilder.alarm: Move to apps/opibuilder
This is a fragment that allows opibuilder scripts to access alarms.
It may only be used in KEK products.
As an alternative to making it an opibuilder fragment,
another way to solve the access problem may be via
Eclipse-RegisterBuddy: org.python.jython
but leaving that unchanged because unsure about its exact usage.