databrowser: Tool bar button for new Data Browser does just that
.. and no longer changing the perspective.
Changing the perspective, i.e. the layout of *.opi views and other
views, needs to be a conscientious user decision and not a side effect
of opening another data browser plot.

This was already done in commit eb041a73f4526631f0ee86d79c0d9609652b7de7
which got lost in a recent merge conflict resolution.