o.c.opibuilder: simple behaviour when OPIView dragged out
There is logic to resize a view when an OPIView is dragged out
of a workbench.  This causes problems when a perspective is
reset or the workbench is restarted.  This commit changes the
behaviour so that the view is located and resized only when
launched from a button.  Otherwise, it locates where put and
the size and shape is the same as it was in the workbench.