Restructure the PerspectiveSaver.
There was previously one small bug with regard to saving OPI layouts.
The perspective was only fully saved to the perspective registry when
also being saved to file.  If you saved a perspective then selected
Window | New Window, OPIs would not be restored unless the perspective
had also been saved to file.

This restructure means that the perspective is always saved to the
registry.