Fix bug where autoscale would sometimes run after dragging
This was being caused as the laxySetAutoScale method changed the state on
the axis temoporarily to create the undoable action. The state was then
queried during this temporary period and an autoscale would run after the
using had finished dragging the plot. To workaround this issue the
autoscale is now provided explicitly to the ChangeAxisRange constructor,
and the helper constructors have been updated accordingly.