Swap order of axis range setting and autoscale setting
If a user checks the autoscale box, then undoes this, the axis ranges were
being set to the previous range while autoscale was still on. They flicked
back to the autoscale range instead of going back to the range they were
set to before the auto-scale box was checked. This fixes this.