Change size of list when creating undoable action
An undoable action is created when panning on a single axis. In this
scenario only one axis is being modified, so the list of new axis auto-scale
values should only contain a single entry. Here it contained two entries,
which raised an InvalidArgument exception which prevented the pan action
from completing.