Medusa removed in favour of a new plug-in inside cs-studio-thirdparty
Due to the many changes I’ve made to the Medusa library and the time
often required before their integration into the master branch and
publishing to Maven central, I’ve opted to a new plug-in into
cs-studio-thirdparty.
This will ease the transition from Medusa to a different project I’m
about to open (based on Medusa code).