Copy StringSplitter.java to org.csstudio.opibuilder.converter.
This breaks the golden rule of avoiding code duplication, but it allows
building the EDM converter into a stand-alone jar file for external use. This
had previously been possible by including the org.csstudio.java plugin, but
since that developed a dependency on an Eclipse plugin that mechanism for
building proved to be tricky as well.