Fix problems with hide toolbar option.
When a plot file specified the hide toolbar option, it would still be
displayed when the plot file was loaded with the data browser OPI widget
and that widget was placed inside a container (e.g. a grouping container
or a tab).

The reason for this was that the code hiding the toolbar incorrectly
checked whether it was visible, so that it would not hide it if it was
set to visible but the container containing the data browser was not
visible yet.