Show PV disconnections in the data browser
Modify the Archiver Appliance reader to set a
special value when a disconnection is found, for
both optimized and raw data.

This appears as a NaN in the data browser. The
data browser is modified to draw a vertical dashed
line to highlight any NaN values.

To distinguish between an archiver disconnect and
other types of NaN value, one may use Inspect Samples.