More specific fix for intermittent background shading problems.
The grey background shading was unreliable for times in the past. It was also unreliable in
the future, but no-one looks in the future in the databrowser.

Painting rectangles with large dimensions can result in an overflow that means they
are painted in the wrong position. To avoid this, check more specifically whether
the grey background is required before painting.