Window focus behaviour for standalone windows.
CS-Studio gives focus to the main window whenever a file is opened.
Usually this is appropriate because something is being opened in
that main window.

However, in the case that a standalone window is being opened, we
don't want to force focus on the main window. This change filters
out that event.